Initial commit: Seanchló keyboard
Some basics here.

This keyboard is designed for the ClóGaelach keymapping in Linux. It'll
probably work elsewhere, but YMMV.

You'll notice the insular forms of r/s. Those are just on the keyboard.
Unless you've modified your keymap, obviously.

We also use the tironian et for ampersand, but that's me being cheeky.

I'm also keeping fadas out of repo name and file names because it messes
up on github. Not like we've had Unicode for 20+ years now or anything.

Signed-off-by: Eilís Ní Fhlannagáin <elizabeth.flanagan@huawei.com>